requirements.txt
pdfminer.six
fastapi
uvicorn[standard]
aiohttp
pydantic-settings
sqlalchemy
asyncmy
jinja2
python-multipart
requests
chromadb